05-engineering/pyproject.toml
[project]
name = "geopolitical-analogue-engine"
version = "0.1.0"
description = "Self-improving geopolitical prediction agent via historical analogue retrieval"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "chromadb>=0.5",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "tqdm>=4.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "matplotlib>=3.0",
    "voyageai>=0.3",
    "datasets>=2.0",
    "openai>=1.0",
    "yfinance>=0.2",
    "jupyter>=1.0",
    "ipykernel>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
